Renamed `lib/` to `src/`.
