+++ packages/typedoc-plugins/rollup.config.js
@@ -25,7 +25,7 @@ const externals = [
 ];
 
 export default defineConfig( {
-	input: 'lib/index.ts',
+	input: 'src/index.ts',
 	output: {
 		format: 'esm',
 		file: path.join( cwd, 'dist', 'index.js' ),
@@ -38,8 +38,6 @@ export default defineConfig( {
 			extensions: [ '.mjs', '.js', '.json', '.node', '.ts', '.mts' ],
 			preferBuiltins: true
 		} ),
-		typescript( {
-			rootDir: './lib'
-		} )
+		typescript()
 	]
 } );
