+++ packages/typedoc-plugins/src/event-inheritance-fixer/index.ts
