+++ packages/typedoc-plugins/src/event-param-fixer/index.ts
