+++ packages/typedoc-plugins/src/index.ts
