+++ packages/typedoc-plugins/src/interface-augmentation-fixer/index.ts
