+++ packages/typedoc-plugins/src/module-fixer/index.ts
