+++ packages/typedoc-plugins/src/purge-private-api-docs/index.ts
