+++ packages/typedoc-plugins/src/symbol-fixer/index.ts
