+++ packages/typedoc-plugins/src/tag-error/augmentation.ts
