+++ packages/typedoc-plugins/src/tag-error/errortagserializer.ts
