+++ packages/typedoc-plugins/src/tag-event/augmentation.ts
