+++ packages/typedoc-plugins/src/tag-event/eventtagserializer.ts
