+++ packages/typedoc-plugins/src/utils/isabsoluteidentifier.ts
