+++ packages/typedoc-plugins/src/utils/isidentifiervalid.ts
