+++ packages/typedoc-plugins/src/validators/fires-validator/index.ts
