+++ packages/typedoc-plugins/src/validators/link-validator/index.ts
