+++ packages/typedoc-plugins/tests/symbol-fixer/index.ts
@@ -9,7 +9,7 @@ import upath from 'upath';
 import { Application, type ProjectReflection, ReflectionKind } from 'typedoc';
 
 import { ROOT_TEST_DIRECTORY } from '../utils.js';
-import { typeDocSymbolFixer } from '../../lib/index.js';
+import { typeDocSymbolFixer } from '../../src/index.js';
 
 describe( 'typedoc-plugins/symbol-fixer', () => {
 	let typeDoc: Application,
